requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
openai==1.72.0
requests==2.32.4
pydantic==2.11.3
pydantic-settings==2.8.1
httpx==0.28.1
python-dotenv==1.1.0
pillow==11.1.0
python-multipart==0.0.22
uuid==1.30
colorama==0.4.6
xai_sdk
#python-jose==3.4.0
#passlib==1.7.4
#bcrypt==4.3.0